requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.5.0
pyyaml==6.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
